@@ -52,122 +52,113 @@
     "heroku-postbuild": "yarn prod"
   },
   "dependencies": {
-    "@salesforce-ux/design-system": "2.17.5",
-    "@salesforce/design-system-react": "0.10.46",
-    "@sentry/browser": "^6.19.2",
+    "@react-hook/window-scroll": "^1.3.0",
+    "@salesforce-ux/design-system": "^2.18.0",
+    "@salesforce/design-system-react": "^0.10.48",
+    "@sentry/browser": "^7.1.1",
     "classnames": "^2.3.1",
-    "date-fns": "^2.27.0",
-    "i18next": "^21.6.14",
+    "date-fns": "^2.28.0",
+    "i18next": "^21.8.9",
     "i18next-browser-languagedetector": "^6.1.4",
-    "i18next-http-backend": "^1.4.0",
+    "i18next-http-backend": "^1.4.1",
     "js-cookie": "^3.0.1",
-    "lodash": "^4.17.20",
-    "react": "^17.0.2",
+    "lodash": "^4.17.21",
+    "react": "^18.1.0",
     "react-document-title": "^2.0.3",
-    "react-dom": "^17.0.2",
-    "react-fns": "^1.4.0",
-    "react-i18next": "^11.16.2",
-    "react-redux": "^7.2.6",
-    "react-router": "^5.3.3",
+    "react-dom": "^18.1.0",
+    "react-i18next": "^11.17.1",
+    "react-redux": "^8.0.2",
     "react-router-dom": "^5.3.3",
-    "redux": "^4.1.2",
-    "redux-devtools-extension": "^2.13.8",
+    "redux": "^4.2.0",
+    "redux-devtools-extension": "^2.13.9",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.4.1",
-    "reselect": "^4.1.5",
+    "reselect": "^4.1.6",
     "sockette": "^2.0.6",
-    "uuid": "^8.3.2",
-    "whatwg-fetch": "^3.5.0"
+    "uuid-random": "^1.3.2",
+    "whatwg-fetch": "^3.6.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.17.8",
-    "@babel/preset-env": "^7.16.11",
-    "@babel/preset-react": "^7.16.7",
-    "@babel/preset-typescript": "^7.16.7",
-    "@storybook/addon-actions": "6.4.19",
-    "@storybook/addon-essentials": "6.4.19",
-    "@storybook/addon-links": "6.4.19",
+    "@babel/core": "^7.18.2",
+    "@babel/preset-env": "^7.18.2",
+    "@babel/preset-react": "^7.17.12",
+    "@babel/preset-typescript": "^7.17.12",
+    "@storybook/addon-actions": "^6.5.8",
+    "@storybook/addon-essentials": "^6.5.8",
+    "@storybook/addon-links": "^6.5.8",
     "@storybook/addon-postcss": "^2.0.0",
-    "@storybook/builder-webpack5": "6.4.19",
-    "@storybook/manager-webpack5": "6.4.19",
-    "@storybook/react": "6.4.19",
+    "@storybook/builder-webpack5": "^6.5.8",
+    "@storybook/manager-webpack5": "^6.5.8",
+    "@storybook/react": "^6.5.8",
     "@storybook/storybook-deployer": "^2.8.11",
-    "@testing-library/jest-dom": "^5.16.3",
-    "@testing-library/react": "^12.1.4",
-    "@testing-library/react-hooks": "^7.0.2",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
     "@types/classnames": "^2.3.1",
     "@types/core-js": "^2.5.5",
-    "@types/js-cookie": "^3.0.1",
-    "@types/lodash": "^4.14.180",
-    "@types/react": "^17.0.43",
+    "@types/js-cookie": "^3.0.2",
+    "@types/lodash": "^4.14.182",
+    "@types/react": "^18.0.12",
     "@types/react-document-title": "^2.0.5",
-    "@types/react-dom": "^17.0.14",
-    "@types/react-redux": "^7.1.23",
+    "@types/react-dom": "^18.0.5",
     "@types/react-router-dom": "^5.3.3",
     "@types/redux-logger": "^3.0.9",
     "@types/redux-mock-store": "^1.0.3",
-    "@types/uuid": "^8.3.4",
-    "@typescript-eslint/eslint-plugin": "^5.16.0",
-    "@typescript-eslint/parser": "^5.16.0",
-    "autoprefixer": "^10.4.4",
-    "babel-jest": "^27.5.1",
-    "babel-loader": "^8.2.4",
-    "core-js": "^3.21.1",
+    "@typescript-eslint/eslint-plugin": "^5.27.1",
+    "@typescript-eslint/parser": "^5.27.1",
+    "autoprefixer": "^10.4.7",
+    "babel-jest": "^28.1.1",
+    "babel-loader": "^8.2.5",
+    "core-js": "^3.22.8",
     "coveralls": "^3.1.1",
     "css-loader": "^6.7.1",
-    "css-minimizer-webpack-plugin": "^3.3.1",
-    "eslint": "^8.11.0",
+    "css-minimizer-webpack-plugin": "^4.0.0",
+    "eslint": "^8.17.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-import-resolver-typescript": "^2.7.0",
+    "eslint-import-resolver-typescript": "^2.7.1",
     "eslint-import-resolver-webpack": "^0.13.2",
-    "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-jest": "^26.1.3",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest": "^26.5.3",
     "eslint-plugin-mdx": "^1.17.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "7.29.4",
-    "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint-plugin-react": "^7.30.0",
+    "eslint-plugin-react-hooks": "^4.5.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
-    "eslint-plugin-storybook": "^0.5.7",
+    "eslint-plugin-storybook": "^0.5.12",
     "fetch-mock": "^9.11.0",
     "html-webpack-plugin": "^5.5.0",
     "i18next-scanner-webpack": "0.8.2",
     "isomorphic-fetch": "^3.0.0",
-    "jest": "^27.5.1",
-    "jest-junit": "^13.0.0",
+    "jest": "^28.1.1",
+    "jest-environment-jsdom": "^28.1.1",
+    "jest-junit": "^13.2.0",
     "mini-css-extract-plugin": "^2.6.0",
     "node-fetch": "^2.6.7",
     "npm-run-all": "^4.1.5",
     "nyc": "^15.1.0",
-    "postcss": "^8.4.12",
-    "postcss-loader": "^6.2.1",
-    "prettier": "^2.6.1",
-    "react-test-renderer": "^17.0.2",
+    "postcss": "^8.4.14",
+    "postcss-loader": "^7.0.0",
+    "prettier": "^2.6.2",
     "redux-mock-store": "^1.5.4",
-    "sass": "^1.49.9",
-    "sass-loader": "^12.6.0",
-    "stylelint": "^14.6.1",
+    "sass": "^1.52.3",
+    "sass-loader": "^13.0.0",
+    "stylelint": "^14.9.0",
     "stylelint-config-prettier": "^9.0.3",
-    "stylelint-config-recommended": "^6.0.0",
-    "stylelint-config-standard-scss": "^3.0.0",
+    "stylelint-config-recommended": "^8.0.0",
+    "stylelint-config-standard-scss": "^4.0.0",
     "stylelint-prettier": "^2.0.0",
-    "typescript": "^4.6.3",
-    "webpack": "^5.70.0",
+    "typescript": "^4.7.3",
+    "webpack": "^5.73.0",
     "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.7.4",
+    "webpack-dev-server": "^4.9.2",
     "webpack-merge": "^5.8.0"
   },
   "resolutions": {
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.0",
     "@storybook/**/ansi-regex": "^5.0.1",
-    "@storybook/**/browserslist": "^4.16.5",
     "@storybook/**/glob-parent": "^5.1.2",
-    "@storybook/**/immer": "^9.0.6",
-    "@storybook/**/prompts": "^2.4.2",
-    "@storybook/**/set-value": "^4.0.1",
-    "@storybook/**/shell-quote": "^1.7.3",
+    "@storybook/**/trim-newlines": "^3.0.1",
     "i18next-scanner-webpack/**/glob-parent": "^5.1.2",
     "trim": "^0.0.3",
-    "shelljs": "^0.8.5"
+    "unset-value": "^2.0.1"
   },
   "yarn-upgrade-all": {
     "ignore": [
