@@ -173,7 +173,7 @@ module.exports = {
 
   // An array of regexp pattern strings that are matched against all source file paths, matched files will skip transformation
   transformIgnorePatterns: [
-    'node_modules/(?!(@salesforce/design-system-react)/)',
+    'node_modules/(?!(@salesforce/design-system-react|@react-hook)/)',
   ],
 
   // An array of regexp pattern strings that are matched against all modules before the module loader will automatically return a mock for them
