@@ -1,4 +1,4 @@
-import { ComponentType } from 'react';
+import { ComponentType, PropsWithChildren } from 'react';
 
-export let Component: ComponentType<any>;
+export let Component: ComponentType<PropsWithChildren<any>>;
 export default Component;
