@@ -1,6 +1,6 @@
-import { t } from 'i18next';
-import * as React from 'react';
+import React, { ReactNode } from 'react';
 import DocumentTitle from 'react-document-title';
+import { useTranslation } from 'react-i18next';
 import { Trans } from 'react-i18next';
 import { Link, RouteComponentProps, withRouter } from 'react-router-dom';
 
@@ -8,11 +8,7 @@ import desertSvg from '@/img/desert.svg?raw';
 import Header from '@/js/components/header';
 import routes from '@/js/utils/routes';
 
-export const EmptyIllustration = ({
-  message,
-}: {
-  message: React.ReactNode;
-}) => (
+export const EmptyIllustration = ({ message }: { message: ReactNode }) => (
   <div className="slds-illustration slds-illustration_large">
     <div
       className="slds-m-vertical_xx-large"
@@ -29,25 +25,29 @@ const FourOhFour = ({
   message,
   history,
 }: {
-  message?: React.ReactNode;
-} & RouteComponentProps) => (
-  <DocumentTitle title={`${t('404')} | ${window.SITE_NAME}`}>
-    <>
-      <Header history={history} />
-      <EmptyIllustration
-        message={
-          message === undefined ? (
-            <Trans i18nKey="pageCannotBeFound">
-              That page cannot be found. Try the{' '}
-              <Link to={routes.home()}>home page</Link>?
-            </Trans>
-          ) : (
-            message
-          )
-        }
-      />
-    </>
-  </DocumentTitle>
-);
+  message?: ReactNode;
+} & RouteComponentProps) => {
+  const { t } = useTranslation();
+
+  return (
+    <DocumentTitle title={`${t('404')} | ${window.SITE_NAME}`}>
+      <>
+        <Header history={history} />
+        <EmptyIllustration
+          message={
+            message === undefined ? (
+              <Trans i18nKey="pageCannotBeFound">
+                That page cannot be found. Try the{' '}
+                <Link to={routes.home()}>home page</Link>?
+              </Trans>
+            ) : (
+              message
+            )
+          }
+        />
+      </>
+    </DocumentTitle>
+  );
+};
 
 export default withRouter(FourOhFour);
