@@ -1,7 +1,7 @@
 import Alert from '@salesforce/design-system-react/components/alert';
 import AlertContainer from '@salesforce/design-system-react/components/alert/container';
-import { t } from 'i18next';
-import * as React from 'react';
+import React, { Component } from 'react';
+import { WithTranslation, withTranslation } from 'react-i18next';
 import { RouteComponentProps } from 'react-router-dom';
 
 import { CurrentJob } from '@/js/store/org/reducer';
@@ -11,9 +11,9 @@ import routes from '@/js/utils/routes';
 type Props = {
   currentJob: CurrentJob;
   history?: RouteComponentProps['history'];
-};
+} & WithTranslation;
 
-class CurrentJobAlert extends React.Component<Props> {
+class CurrentJobAlert extends Component<Props> {
   redirectToJob = () => {
     const { currentJob, history } = this.props;
     const { product_slug, version_label, plan_slug, id } = currentJob;
@@ -26,7 +26,7 @@ class CurrentJobAlert extends React.Component<Props> {
   };
 
   render() {
-    const { currentJob } = this.props;
+    const { t, currentJob } = this.props;
     const { plan_average_duration } = currentJob;
     const duration = getDuration(plan_average_duration);
     let heading = t('An installation is currently running on this org.');
@@ -49,4 +49,4 @@ class CurrentJobAlert extends React.Component<Props> {
   }
 }
 
-export default CurrentJobAlert;
+export default withTranslation()(CurrentJobAlert);
