@@ -1,5 +1,6 @@
 import SLDSProgressIndicator from '@salesforce/design-system-react/components/progress-indicator';
-import * as React from 'react';
+import React from 'react';
+import { useTranslation } from 'react-i18next';
 
 import { getSteps } from '@/js/components/plans/progressIndicator';
 import { Job } from '@/js/store/jobs/reducer';
@@ -15,7 +16,9 @@ const ProgressIndicator = ({
   supportedOrgs: SupportedOrgs;
   preflightRequired: boolean;
 }) => {
-  const steps = getSteps(supportedOrgs, preflightRequired);
+  const { t } = useTranslation();
+
+  const steps = getSteps(t, supportedOrgs, preflightRequired);
   const completedSteps =
     job.status === CONSTANTS.STATUS.COMPLETE
       ? steps.slice()
