@@ -1,7 +1,7 @@
 import Button from '@salesforce/design-system-react/components/button';
 import Spinner from '@salesforce/design-system-react/components/spinner';
-import { t } from 'i18next';
-import * as React from 'react';
+import React, { Component, ReactNode } from 'react';
+import { WithTranslation, withTranslation } from 'react-i18next';
 import { RouteComponentProps } from 'react-router-dom';
 
 import Login from '@/js/components/header/login';
@@ -43,7 +43,7 @@ type Props = {
     email: string,
   ) => Promise<ScratchOrgSpinning>;
   doLogout: () => Promise<FetchOrgJobsSucceeded>;
-};
+} & WithTranslation;
 
 const { AUTO_START_PREFLIGHT, RESULT_STATUS, STATUS } = CONSTANTS;
 const btnClasses = 'slds-p-vertical_xx-small';
@@ -96,7 +96,7 @@ export const ActionBtn = ({
   btnVariant,
   className,
 }: {
-  label: string | React.ReactNode;
+  label: string | ReactNode;
   disabled?: boolean;
   onClick?: () => void;
   btnVariant?: string;
@@ -111,7 +111,7 @@ export const ActionBtn = ({
   />
 );
 
-class CtaButton extends React.Component<
+class CtaButton extends Component<
   Props,
   {
     preflightModalOpen: boolean;
@@ -317,7 +317,7 @@ class CtaButton extends React.Component<
   }
 
   getPersistentOrgCTA() {
-    const { user, clickThroughAgreement, plan, preflight, selectedSteps } =
+    const { t, user, clickThroughAgreement, plan, preflight, selectedSteps } =
       this.props;
     const { preflightModalOpen } = this.state;
     const usesBothOrgTypes = plan.supported_orgs === SUPPORTED_ORGS.Both;
@@ -447,6 +447,7 @@ class CtaButton extends React.Component<
 
   getScratchOrgCTA() {
     const {
+      t,
       user,
       clickThroughAgreement,
       plan,
@@ -626,7 +627,7 @@ class CtaButton extends React.Component<
   }
 
   render() {
-    const { plan, preflight, scratchOrg } = this.props;
+    const { t, plan, preflight, scratchOrg } = this.props;
     const canUsePersistentOrg = plan.supported_orgs !== SUPPORTED_ORGS.Scratch;
     const canUseScratchOrg = Boolean(
       window.GLOBALS.SCRATCH_ORGS_AVAILABLE &&
@@ -675,13 +676,17 @@ class CtaButton extends React.Component<
 
     return (
       <>
-        {canUsePersistentOrg && !hasValidScratchOrg
-          ? this.getPersistentOrgCTA()
-          : null}
-        {canUseScratchOrg ? this.getScratchOrgCTA() : null}
+        {canUsePersistentOrg && !hasValidScratchOrg ? (
+          <div className="slds-m-bottom_small">
+            {this.getPersistentOrgCTA()}
+          </div>
+        ) : null}
+        {canUseScratchOrg ? (
+          <div className="slds-m-bottom_small">{this.getScratchOrgCTA()}</div>
+        ) : null}
       </>
     );
   }
 }
 
-export default CtaButton;
+export default withTranslation()(CtaButton);
