@@ -1,5 +1,5 @@
 import PageHeader from '@salesforce/design-system-react/components/page-header';
-import * as React from 'react';
+import React, { ReactNode } from 'react';
 import { Trans } from 'react-i18next';
 import { Link } from 'react-router-dom';
 
@@ -27,7 +27,7 @@ const Header = ({
   product: Product;
   version: Version;
   plan: Plan;
-  onRenderActions?: () => React.ReactNode;
+  onRenderActions?: () => ReactNode;
   job?: Job;
   userLoggedIn?: boolean;
   scratchOrgCreated?: boolean;
