@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 import { Trans } from 'react-i18next';
 import { Link } from 'react-router-dom';
 
