@@ -1,7 +1,7 @@
 import DataTableCell from '@salesforce/design-system-react/components/data-table/cell';
 import classNames from 'classnames';
-import { t } from 'i18next';
-import * as React from 'react';
+import React from 'react';
+import { useTranslation } from 'react-i18next';
 
 import { DataCellProps } from '@/js/components/plans/stepsTable';
 import { CONSTANTS } from '@/js/store/plans/reducer';
@@ -9,6 +9,7 @@ import { CONSTANTS } from '@/js/store/plans/reducer';
 const { RESULT_STATUS } = CONSTANTS;
 
 const RequiredDataCell = (props: DataCellProps) => {
+  const { t } = useTranslation();
   const { preflight, item, job, className, activeJobStep, ...otherProps } =
     props;
 
