@@ -1,8 +1,8 @@
 import Card from '@salesforce/design-system-react/components/card';
 import Icon from '@salesforce/design-system-react/components/icon';
 import { format, parseISO } from 'date-fns';
-import { t } from 'i18next';
-import * as React from 'react';
+import React from 'react';
+import { useTranslation } from 'react-i18next';
 import { Trans } from 'react-i18next';
 
 import noConnectionSvg from '@/img/no-connection.svg?raw';
@@ -13,29 +13,37 @@ import { ScratchOrg } from '@/js/store/scratchOrgs/reducer';
 import { User } from '@/js/store/user/reducer';
 import { SUPPORTED_ORGS } from '@/js/utils/constants';
 
-const LoggedOut = () => (
-  <div className="slds-illustration slds-illustration_small">
-    <div
-      className="slds-m-vertical_medium"
-      dangerouslySetInnerHTML={{ __html: noConnectionSvg }}
-    />
-    <h3 className="slds-illustration__header slds-text-heading_medium">
-      {t('Not Connected to Salesforce')}
-    </h3>
-  </div>
-);
+const LoggedOut = () => {
+  const { t } = useTranslation();
 
-const Footer = () => (
-  <Trans i18nKey="switchOrg">
-    Is this the correct org? If not, please{' '}
-    <Login
-      id="user-info-login"
-      label={t('log in with a different org')}
-      buttonClassName="slds-p-horizontal_xxx-small"
-      buttonVariant="base"
-    />
-  </Trans>
-);
+  return (
+    <div className="slds-illustration slds-illustration_small">
+      <div
+        className="slds-m-vertical_medium"
+        dangerouslySetInnerHTML={{ __html: noConnectionSvg }}
+      />
+      <h3 className="slds-illustration__header slds-text-heading_medium">
+        {t('Not Connected to Salesforce')}
+      </h3>
+    </div>
+  );
+};
+
+const Footer = () => {
+  const { t } = useTranslation();
+
+  return (
+    <Trans i18nKey="switchOrg">
+      Is this the correct org? If not, please{' '}
+      <Login
+        id="user-info-login"
+        label={t('log in with a different org')}
+        buttonClassName="slds-p-horizontal_xxx-small"
+        buttonVariant="base"
+      />
+    </Trans>
+  );
+};
 
 const UserInfo = ({
   user,
@@ -46,6 +54,8 @@ const UserInfo = ({
   plan: Plan;
   scratchOrg?: ScratchOrg | null;
 }) => {
+  const { t } = useTranslation();
+
   const hasValidToken = Boolean(user?.valid_token_for);
   const username = user?.username;
   const org_name = user?.org_name;
