@@ -1,6 +1,6 @@
 import PageHeader from '@salesforce/design-system-react/components/page-header';
-import { t } from 'i18next';
-import * as React from 'react';
+import React from 'react';
+import { useTranslation } from 'react-i18next';
 import { Trans } from 'react-i18next';
 
 import ProductIcon from '@/js/components/products/icon';
@@ -13,26 +13,30 @@ const Header = ({
 }: {
   product: Product;
   versionLabel: string;
-}) => (
-  <PageHeader
-    className="page-header slds-p-around_x-large"
-    title={
-      product.layout === PRODUCT_LAYOUTS.Card
-        ? product.title
-        : t('Select a Plan')
-    }
-    trail={
-      product.layout === PRODUCT_LAYOUTS.Card
-        ? []
-        : [
-            <Trans i18nKey="productWithVersion" key={product.slug}>
-              {{ product: product.title }} {{ version: versionLabel }}
-            </Trans>,
-          ]
-    }
-    icon={<ProductIcon item={product} />}
-    variant="object-home"
-  />
-);
+}) => {
+  const { t } = useTranslation();
+
+  return (
+    <PageHeader
+      className="page-header slds-p-around_x-large"
+      title={
+        product.layout === PRODUCT_LAYOUTS.Card
+          ? product.title
+          : t('Select a Plan')
+      }
+      trail={
+        product.layout === PRODUCT_LAYOUTS.Card
+          ? []
+          : [
+              <Trans i18nKey="productWithVersion" key={product.slug}>
+                {{ product: product.title }} {{ version: versionLabel }}
+              </Trans>,
+            ]
+      }
+      icon={<ProductIcon item={product} />}
+      variant="object-home"
+    />
+  );
+};
 
 export default Header;
