@@ -1,6 +1,6 @@
 import Avatar from '@salesforce/design-system-react/components/avatar';
 import Icon from '@salesforce/design-system-react/components/icon';
-import * as React from 'react';
+import React, { CSSProperties } from 'react';
 
 import { Product } from '@/js/store/products/reducer';
 
@@ -33,7 +33,7 @@ const ProductIcon = ({ item }: { item: Product }) => {
   if (item.color) {
     // Standard entity icon (initials) with custom color
     return (
-      <div style={{ '--custom-color': item.color } as React.CSSProperties}>
+      <div style={{ '--custom-color': item.color } as CSSProperties}>
         <Avatar variant="entity" label={item.title} />
       </div>
     );
