@@ -1,5 +1,5 @@
-import { t } from 'i18next';
-import * as React from 'react';
+import React from 'react';
+import { useTranslation } from 'react-i18next';
 import { Trans } from 'react-i18next';
 
 import svgPath from '@/img/no-access.svg?raw';
@@ -17,43 +17,49 @@ const NotAllowed = ({
   isLoggedIn: boolean;
   message: string | null;
   link: JSX.Element;
-}) => (
-  <>
-    <div className="slds-illustration slds-illustration_large">
-      <div
-        className="slds-m-vertical_xx-large"
-        dangerouslySetInnerHTML={{ __html: svgPath }}
-      />
-      <h3 className="slds-illustration__header slds-text-heading_medium">
-        {t('Restricted Access')}
-      </h3>
-    </div>
-    {message ? (
-      <div className="slds-align_absolute-center slds-size_2-of-3">
+}) => {
+  const { t } = useTranslation();
+
+  return (
+    <>
+      <div className="slds-illustration slds-illustration_large">
         <div
-          className="slds-text-longform slds-text-body_regular markdown"
-          // This message is pre-cleaned by the API
-          dangerouslySetInnerHTML={{
-            __html: message,
-          }}
+          className="slds-m-vertical_xx-large"
+          dangerouslySetInnerHTML={{ __html: svgPath }}
         />
+        <h3 className="slds-illustration__header slds-text-heading_medium">
+          {t('Restricted Access')}
+        </h3>
       </div>
-    ) : null}
-    <div className="slds-align_absolute-center">
-      <div className="slds-text-longform slds-text-body_regular">
-        <Trans i18nKey="a_or_b">
-          <NotAllowedLink link={link} />
-          {' or '}
-          <Login
-            id="product-not-allowed-login"
-            label={isLoggedIn ? t('log in with a different org') : t('log in')}
-            buttonClassName="slds-p-horizontal_xxx-small"
-            buttonVariant="base"
+      {message ? (
+        <div className="slds-align_absolute-center slds-size_2-of-3">
+          <div
+            className="slds-text-longform slds-text-body_regular markdown"
+            // This message is pre-cleaned by the API
+            dangerouslySetInnerHTML={{
+              __html: message,
+            }}
           />
-        </Trans>
+        </div>
+      ) : null}
+      <div className="slds-align_absolute-center">
+        <div className="slds-text-longform slds-text-body_regular">
+          <Trans i18nKey="a_or_b">
+            <NotAllowedLink link={link} />
+            {' or '}
+            <Login
+              id="product-not-allowed-login"
+              label={
+                isLoggedIn ? t('log in with a different org') : t('log in')
+              }
+              buttonClassName="slds-p-horizontal_xxx-small"
+              buttonVariant="base"
+            />
+          </Trans>
+        </div>
       </div>
-    </div>
-  </>
-);
+    </>
+  );
+};
 
 export default NotAllowed;
