@@ -1,6 +1,6 @@
 import Card from '@salesforce/design-system-react/components/card';
-import { t } from 'i18next';
-import * as React from 'react';
+import React from 'react';
+import { useTranslation } from 'react-i18next';
 import { Link } from 'react-router-dom';
 
 import { Plan } from '@/js/store/plans/reducer';
@@ -16,32 +16,35 @@ const PlanCard = ({
   productSlug: string;
   versionLabel: string;
   plan: Plan;
-}) => (
-  <div
-    className="slds-p-around_medium
-      slds-size_1-of-1
-      slds-medium-size_1-of-2
-      slds-large-size_1-of-3"
-  >
-    <Card heading={plan.title} bodyClassName="slds-card__body_inner">
-      {plan.preflight_message ? (
-        <div
-          className="markdown"
-          // These messages are pre-cleaned by the API
-          dangerouslySetInnerHTML={{
-            __html: plan.preflight_message,
-          }}
-        />
-      ) : null}
-      <Link
-        to={routes.plan_detail(productSlug, versionLabel, plan.slug)}
-        className="slds-button slds-button_brand"
-      >
-        {t('View Details')}
-      </Link>
-    </Card>
-  </div>
-);
+}) => {
+  const { t } = useTranslation();
+  return (
+    <div
+      className="slds-p-around_medium
+        slds-size_1-of-1
+        slds-medium-size_1-of-2
+        slds-large-size_1-of-3"
+    >
+      <Card heading={plan.title} bodyClassName="slds-card__body_inner">
+        {plan.preflight_message ? (
+          <div
+            className="markdown"
+            // These messages are pre-cleaned by the API
+            dangerouslySetInnerHTML={{
+              __html: plan.preflight_message,
+            }}
+          />
+        ) : null}
+        <Link
+          to={routes.plan_detail(productSlug, versionLabel, plan.slug)}
+          className="slds-button slds-button_brand"
+        >
+          {t('View Details')}
+        </Link>
+      </Card>
+    </div>
+  );
+};
 
 const PlanCards = ({
   product,
